Split races_admin router into a package by concern; add missing admin endpoint tests
routers/races_admin.py (~870 lines) is now routers/races_admin/ with helpers.py
(shared reconciliation logic), records.py (race CRUD, status recheck, run
trigger), drafts.py (delete/publish/unpublish/batch-publish), forecasts.py
(chamber-forecast draft/generate/publish), and race_runs.py/data.py for
per-race run history and full race JSON reads. The combined router is
assembled in __init__.py in the same relative order as the original file so
the one order-sensitive route pair (chamber_forecasts/publish vs the
parameterized {race_id}/publish) still resolves the same way.

Route parity (path, method, dependency) verified programmatically against the
original single-file module: all 19 endpoints match exactly, and a live
TestClient request confirms /api/races/chamber_forecasts/publish is not
shadowed by /api/races/{race_id}/publish.

Also added tests/test_races_api_admin.py coverage for five admin endpoints
that had no test at all: POST .../cancel, DELETE .../draft, GET .../runs,
GET .../runs/{run_id}, and GET .../data.
